Frontent/app/backend/requirements.txt
azure-identity==1.13.0b3
Flask==2.2.5
langchain==0.0.187
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
black==23.7.0
flake8==6.0.0
pytest==7.4.0
python-dotenv==1.0.0
azure-ai-translation-text==1.0.0b1
openai==1.3.5
